+++ src/MultipleSelect.js
@@ -91,7 +91,7 @@ class MultipleSelect {
       this.$drop.append(`
         <div class="ms-search">
           <input type="text" autocomplete="off" autocorrect="off"
-            autocapitilize="off" spellcheck="false"
+            autocapitalize="off" spellcheck="false"
             ${sprintf`placeholder="${s}"`(this.options.filterPlaceholder)}>
         </div>
       `)
@@ -156,7 +156,8 @@ class MultipleSelect {
     if ($elm.is('option')) {
       const text = this.options.textTemplate($elm)
       const {value, selected} = el
-      const style = sprintf`style="${s}"`(this.options.styler(value))
+      const customStyle = this.options.styler(value)
+      const style = customStyle ? sprintf`style="${s}"`(customStyle) : ''
 
       disabled = groupDisabled || el.disabled
 
@@ -342,9 +343,10 @@ class MultipleSelect {
         top: offset.top,
         left: offset.left
       })
+      this.$drop.outerWidth(this.$parent.outerWidth())
     }
 
-    if (this.options.filter) {
+    if (this.$el.children().length && this.options.filter) {
       this.$searchInput.val('')
       this.$searchInput.focus()
       this.filter()
@@ -382,26 +384,27 @@ class MultipleSelect {
   }
 
   update (ignoreTrigger) {
-    const selects = this.options.displayValues ? this.getSelects() : this.getSelects('text')
+    const valueSelects = this.getSelects()
+    const textSelects = this.options.displayValues ? valueSelects : this.getSelects('text')
     const $span = this.$choice.find('>span')
-    const sl = selects.length
+    const sl = valueSelects.length
 
     if (sl === 0) {
       $span.addClass('placeholder').html(this.options.placeholder)
     } else if (this.options.formatAllSelected() && sl === this.$selectItems.length + this.$disableItems.length) {
       $span.removeClass('placeholder').html(this.options.formatAllSelected())
     } else if (this.options.ellipsis && sl > this.options.minimumCountSelected) {
-      $span.removeClass('placeholder').text(`${selects.slice(0, this.options.minimumCountSelected)
+      $span.removeClass('placeholder').text(`${textSelects.slice(0, this.options.minimumCountSelected)
         .join(this.options.displayDelimiter)}...`)
     } else if (this.options.formatCountSelected() && sl > this.options.minimumCountSelected) {
       $span.removeClass('placeholder').html(this.options.formatCountSelected()
-        .replace(/#/g, selects.length)
+        .replace(/#/g, sl)
         .replace(/%/g, this.$selectItems.length + this.$disableItems.length))
     } else {
-      $span.removeClass('placeholder').text(selects.join(this.options.displayDelimiter))
+      $span.removeClass('placeholder').text(textSelects.join(this.options.displayDelimiter))
     }
 
-    if (this.options.addTitle) {
+    if (this.options.displayTitle) {
       $span.prop('title', this.getSelects('text'))
     }
 
@@ -558,7 +561,9 @@ class MultipleSelect {
       if (!this.options.filterGroup) {
         this.$selectItems.each((i, el) => {
           const $parent = $(el).parent()
-          $parent[!removeDiacritics($parent.text().toLowerCase()).includes(removeDiacritics(text)) ? 'hide' : 'show']()
+          const hasText = removeDiacritics($parent.text().toLowerCase())
+            .includes(removeDiacritics(text))
+          $parent.closest('li')[hasText ? 'show' : 'hide']()
         })
       }
       this.$disableItems.parent().hide()
@@ -566,12 +571,15 @@ class MultipleSelect {
         const $parent = $(el).parent()
         const group = $parent[0].getAttribute('data-group')
         if (this.options.filterGroup) {
-          const func = !removeDiacritics($parent.text().toLowerCase()).includes(removeDiacritics(text)) ? 'hide' : 'show'
-          $parent[func]()
-          this.$selectItems.filter(`[data-group="${group}"]`).parent()[func]()
+          const hasText = removeDiacritics($parent.text().toLowerCase())
+            .includes(removeDiacritics(text))
+          const func = hasText ? 'show' : 'hide'
+          $parent.closest('li')[func]()
+          this.$selectItems.filter(`[data-group="${group}"]`).closest('li')[func]()
         } else {
           const $items = this.$selectItems.filter(':visible')
-          $parent[$items.filter(sprintf`[data-group="${s}"]`(group)).length ? 'show' : 'hide']()
+          const hasText = $items.filter(sprintf`[data-group="${s}"]`(group)).length
+          $parent.closest('li')[hasText ? 'show' : 'hide']()
         }
       })
 
@@ -610,6 +618,7 @@ const defaults = {
   position: 'bottom',
 
   displayValues: false,
+  displayTitle: false,
   displayDelimiter: ', ',
   minimumCountSelected: 3,
   ellipsis: false,